refactoring approach for login in cypress
