+++ cypress/cypress-smoketests/integration/change-address.cy.ts
@@ -7,19 +7,20 @@ import {
   updateAddress1,
 } from '../support/pages/my-account';
 
-const { getUserToken, login } = getEnvironmentSpecificFunctions();
+const { login } = getEnvironmentSpecificFunctions();
 
 const DEFAULT_ACCOUNT_PASS = Cypress.env('DEFAULT_ACCOUNT_PASS');
 
 describe('Private practitioner', () => {
   let token = null;
 
-  before(async () => {
-    const results = await getUserToken(
-      'privatePractitioner1@example.com',
-      DEFAULT_ACCOUNT_PASS,
-    );
-    token = results.AuthenticationResult.IdToken;
+  before(() => {
+    cy.task('getUserToken', {
+      email: 'privatePractitioner1@example.com',
+      password: DEFAULT_ACCOUNT_PASS,
+    }).then(result => {
+      token = result.AuthenticationResult.IdToken;
+    });
   });
 
   it('logs in', () => {
