+++ cypress/cypress-smoketests/integration/user-post-authentication.cy.ts
@@ -1,5 +1,3 @@
-import { type InitiateAuthResponse } from 'aws-sdk/clients/cognitoidentityserviceprovider';
-import { confirmUser } from '../support/pages/login';
 import {
   editPetitionerEmail,
   goToCaseDetail,
@@ -15,8 +13,7 @@ import {
   serveCaseToIrs,
 } from '../support/pages/create-paper-case';
 import { goToMyDocumentQC } from '../support/pages/document-qc';
-const { closeScannerSetupDialog, getUserToken, login } =
-  getEnvironmentSpecificFunctions();
+const { closeScannerSetupDialog, login } = getEnvironmentSpecificFunctions();
 
 let token: string | undefined = undefined;
 const testData = {};
@@ -26,12 +23,13 @@ const randomizedEmail = `${faker.string.uuid()}@example.com`;
 
 if (!Cypress.env('SMOKETESTS_LOCAL')) {
   describe('Petitions clerk', () => {
-    before(async () => {
-      const results: InitiateAuthResponse = await getUserToken(
-        'petitionsclerk1@example.com',
-        DEFAULT_ACCOUNT_PASS,
-      );
-      token = results.AuthenticationResult?.IdToken;
+    before(() => {
+      cy.task('getUserToken', {
+        email: 'petitionsclerk1@example.com',
+        password: DEFAULT_ACCOUNT_PASS,
+      }).then(result => {
+        token = result.AuthenticationResult?.IdToken;
+      });
     });
 
     it('should be able to login', () => {
@@ -49,12 +47,13 @@ if (!Cypress.env('SMOKETESTS_LOCAL')) {
   });
 
   describe('Admission clerk', () => {
-    before(async () => {
-      const results: InitiateAuthResponse = await getUserToken(
-        'admissionsclerk1@example.com',
-        DEFAULT_ACCOUNT_PASS,
-      );
-      token = results.AuthenticationResult?.IdToken;
+    before(() => {
+      cy.task('getUserToken', {
+        email: 'admissionsclerk1@example.com',
+        password: DEFAULT_ACCOUNT_PASS,
+      }).then(result => {
+        token = result.AuthenticationResult?.IdToken;
+      });
     });
 
     it('should be able to login', () => {
@@ -69,16 +68,17 @@ if (!Cypress.env('SMOKETESTS_LOCAL')) {
   });
 
   describe('Petitioner', () => {
-    it('should confirm user', async () => {
-      await confirmUser({ email: randomizedEmail });
+    it('should confirm user', () => {
+      cy.task('confirmUser', { email: randomizedEmail });
     });
 
-    it('verify the email has changed', async () => {
-      const results: InitiateAuthResponse = await getUserToken(
-        randomizedEmail,
-        DEFAULT_ACCOUNT_PASS,
-      );
-      token = results.AuthenticationResult?.IdToken;
+    it('verify the email has changed', () => {
+      cy.task('getUserToken', {
+        email: randomizedEmail,
+        password: DEFAULT_ACCOUNT_PASS,
+      }).then(result => {
+        token = result.AuthenticationResult?.IdToken;
+      });
       login(token);
       goToCaseDetailPetitioner(testData.createdPaperDocketNumber);
       verifyEmailChange();
